refactor: 수익 계산을 int 로 반환해주도록 리팩토링

+++ src/main/java/lotto/domain/Money.java
@@ -28,8 +28,8 @@ public class Money {
         return money / 1000;
     }
 
-    public float calculateProfitRate(float profit) {
-        return profit / money;
+    public float calculateProfitRate(int profit) {
+        return (float) profit / money;
     }
 
     @Override
